refactor(client): curate public barrels and convert cross-slice imports
