+++ client/src/features/analytics/index.ts
@@ -5,5 +5,9 @@
 // R7: cross-slice imports must go through this barrel only; importing a path
 // inside another slice's internals (features/analytics/pages/..., etc.) is a
 // boundaries lint violation.
-//
-// Empty on purpose: no files have moved into this slice yet.
+
+export { default as Dashboard } from './pages/Dashboard';
+export { default as Exports } from './pages/Exports';
+export { default as ReportBuilder } from './pages/ReportBuilder';
+export { default as AiInsights } from './pages/AiInsights';
+export { default as AdvancedAnalytics } from './pages/AdvancedAnalytics';
