+++ client/src/features/sales/index.ts
@@ -5,5 +5,8 @@
 // R7: cross-slice imports must go through this barrel only; importing a path
 // inside another slice's internals (features/sales/pages/..., etc.) is a
 // boundaries lint violation.
-//
-// Empty on purpose: no files have moved into this slice yet.
+
+export { default as SalesHistory } from './pages/SalesHistory';
+export { default as Promotions } from './pages/Promotions';
+export { default as GiftCards } from './pages/GiftCards';
+export { default as Layaway } from './pages/Layaway';
